nvptx-none-run: do not use CUDA runtime library (GH-2)
The call to cudaDeviceReset does not serve any purpose: this function can be
used for purging resources allocated on the device, but as this is the last
call from main, the process is about to exit and resources will be released
anyway.  The call is present there from the initial commit without a comment.

Handling of CUDA runtime in configure.ac is removed in a followup commit.