nvptx-none-run: extend stack limiting (GH-8)
It appears that memory accounting in CUDA drivers can be very conservative,
sometimes applying stack reservation for the maximum amount of threads the
device can host.  If so, even the current limit of 128 KiB is too much for
some Maxwell devices.  Instead of hardcoding some limit, calculate it at
run time based on available memory and maximum thread count.  Allow the user
to override this automatic guess on the command line.