Ensure :VAR_MAP and :FUNC_MAP are output in order
Before, only if there is an '.extern', is_decl is set and only
then the variable/function is added to 'decls'. However, global
variables/functions may not have .extern - even if they show up
with :VAR_MAP and :FUNC_MAP. GCC's mkoffload uses the latter and
requires that the order matches the original order.

As the output is done as:
  write_stmts (out, rev_stmts (decls));
  htab_traverse (symbol_table, traverse, (void *)out);
  write_stmts (out, rev_stmts (fns));
it might have happen that :VAR_MAP/:FUNC_MAP lines will be output
in hash-map order, which can cause the wrong order.

	https://gcc.gnu.org/PR100059
	* nvptx-as.c (parse_file): Set is_decl when there is a
	:VAR_MAP or :FUNC_MAP.