ld: Fix undefined behavior reading objects from archives
Per 'xstrdup' usage and 'while (*ptx != '\0')' in 'process_refs_defs' (..., and
subsequently applying to 'data' in 'file_hash_entry' instances), the code
currently assumes that 'get_contents ()' of an archive is NUL-terminated.
However, objects in archives generally are not NUL-terminated, and what happens
then is that the 'xstrdup' overruns the buffer until it comes across a NUL.
Instead of 'xstrdup', fix it to the known length, as done elsewhere.