Handle --hash-style argument in nvptx-ld
The gcc build infrastructure passes --hash-style=STYLE to the linker,
including when building the offload compiler.  This fails with
nvptx-ld because it does not understand the option.

The functionality selectd by this option it is not needed but to ease
the use of the tools nvptx-ld could just ignore the option.

This patch does just that.  It could potentially be made more strict
by checking for the exact option argument but there is not really a
reason for that.  Regardless of how the host tools are build, the
nvptx backend won't use the ELF hashing indicated.